requirements.txt
python-telegram-bot>=20.0
requests>=2.28.0
beautifulsoup4>=4.11.0
openai>=1.0.0
python-dotenv>=1.0.0
pytz>=2023.3
